CRUD Hibernate con Jboss tool

+++ CRUDHibernate/src/main/java/com/isotrol/repository/PersonaDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.isotrol.entity.Persona;
 
-public interface PersonaDAO extends JpaRepository<Persona, Long>{
+public interface PersonaDAO extends JpaRepository<Persona, Integer>{
 
 }
